Undo unrelated `BN` import change
This change was required due to an unrelated accidental breaking change
made in `@metamask/controllers` that was included in v6.2.0. The
breaking change was reversed in v6.2.1, so this modification was no
longer required.

It would still be nice to not use this `BN` import though. This is
removed in a separate PR:
https://github.com/MetaMask/metamask-mobile/pull/2431